made triangle the first shared struct
